TwitPicEngine rewritten to use QNetworkAccessManager instead of obsolete QHttp.

+++ qtwitter-app/src/core.h
@@ -90,7 +90,7 @@ signals:
   void accountsUpdated( const QList<Account> &accounts, int isPublicTimelineRequested );
   void errorMessage( const QString &message );
   void twitPicResponseReceived();
-  void twitPicDataSendProgress(int,int);
+  void twitPicDataSendProgress(qint64,qint64);
   void requestListRefresh( bool isPublicTimeline, bool isSwitchUser);
   void requestStarted();
   void resetUi();
